backend/pyproject.toml
[tool.poetry]
name = "edgar"
version = "0.1.0"
description = "Infers types for CSV and excel files using a webapp"
authors = ["Georgea75 <georgetucker.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
pandas = "^2.2.1"
pydantic = "^2.6.3"
djantic = "^0.7.0"
parameterized = "^0.9.0"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
